# controller-model/routes/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "The APIs for zlz - a photo sharing social network - are listed below for your use. For the source code of our project, please see [https://github.com/cis557/cis557-project-photo-sharing-social-network-app-zlz](https://github.com/cis557/cis557-project-photo-sharing-social-network-app-zlz)",
    "version": "1.0.0",
    "title": "zlz - photo sharing social network",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "zzm@seas.upenn.edu"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "peaceful-forest-83015.herokuapp.com/",
...
  "schemes": [
    "https"
  ],
  "paths": {
    "/user/{id}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "READ a user by its id",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "description": "id of the user you want to get",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "404": {
            "description": "fail to get the user"
          }
        }
      }
    },
...
    "/post": {
      "post": {
        "tags": [
          "post"
        ],
        "summary": "CREATE a post",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "author",
            "description": "id of the author",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "title",
            "description": "title of the post",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "image",
            "description": "image of the post as a base64 string",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "description",
            "description": "description of the post",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "tag",
            "description": "tags of the post",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "post created"
          },
          "400": {
            "description": "missing parameters"
          },
          "404": {
            "description": "fail to create the post"
          }
        }
      }
    },
    "/post/{id}": {
      "get": {
        "tags": [
          "post"
        ],
        "summary": "READ a post by its id",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "description": "id of the post you want to get",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "404": {
            "description": "fail to get the post"
          }
        }
      }
    },
    "/post/put": {
      "put": {
        "tags": [
          "post"
        ],
        "summary": "UPDATE a post by its id",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "id",
            "type": "string",
            "description": "id of the post you want to update",
            "required": true
          },
          {
            "in": "body",
            "name": "title",
            "type": "string",
            "description": "updated title of the post",
            "required": true
          },
          {
            "in": "body",
            "name": "description",
            "type": "string",
            "description": "updated description of the post you want to update",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "404": {
            "description": "fail to update the post"
          }
        }
      }
    },
    "/post/delete": {
      "delete": {
        "tags": [
          "post"
        ],
        "summary": "DELETE a post by its id",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "id",
            "type": "string",
            "description": "id of the post you want to delete",
            "required": true
          },
          {
            "in": "body",
            "name": "userId",
            "type": "string",
            "description": "id of the current user",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "404": {
            "description": "fail to delete the post"
          }
        }
      }
    },
...
    "/profile/{id}": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "READ a user by its id",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "description": "id of the user you want to get",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "404": {
            "description": "fail to get the user"
          }
        }
      }
    },
    "/profile/comment/{id}": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "READ comments for user profile by user id",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "description": "id of the user you want to get",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "404": {
            "description": "fail to get the comments for user profile"
          }
        }
      }
    },
    "/profile/like/{id}": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "READ liking information for user profile by user id",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "description": "id of the user you want to get",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "404": {
            "description": "fail to get the liking information for user profile"
          }
        }
      }
    },
    "/profile/follow/{id}": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "READ followship information for user profile by user id",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "description": "id of the user you want to get",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "404": {
            "description": "fail to get the followship information for user profile"
          }
        }
      }
    }
  }
}